% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_constants.R
\docType{data}
\name{.api_GET_specs}
\alias{.api_GET_specs}
\title{GET API Specifications for OKX}
\format{
An object of class \code{list} of length 11.
}
\usage{
.api_GET_specs
}
\description{
Internal list containing GET endpoint configurations.
}
\keyword{(`/api/v5/account/balance`)}
\keyword{(`/api/v5/account/leverage-info`)}
\keyword{(`/api/v5/account/positions-history`)}
\keyword{(`/api/v5/account/positions`)}
\keyword{(`/api/v5/asset/balances`)}
\keyword{(`/api/v5/asset/deposit-history`)}
\keyword{(`/api/v5/asset/withdrawal-history`)}
\keyword{(`/api/v5/market/candles`)}
\keyword{(`/api/v5/market/history-candles`)}
\keyword{(`/api/v5/public/mark-price`)}
\keyword{(`/api/v5/trade/order`)}
\keyword{Fetches}
\keyword{ID.}
\keyword{Retrieves}
\keyword{Specification}
\keyword{a}
\keyword{account}
\keyword{account-level}
\keyword{account.}
\keyword{all}
\keyword{an}
\keyword{and}
\keyword{asset}
\keyword{assets}
\keyword{balance}
\keyword{balances}
\keyword{bar.}
\keyword{before}
\keyword{by}
\keyword{candlestick}
\keyword{candlesticks}
\keyword{current}
\keyword{data}
\keyword{deposit}
\keyword{deposits.}
\keyword{details}
\keyword{details.}
\keyword{equity}
\keyword{for}
\keyword{getting}
\keyword{given}
\keyword{historical}
\keyword{history}
\keyword{in}
\keyword{info}
\keyword{information.}
\keyword{instrument}
\keyword{instrument.}
\keyword{internal}
\keyword{latest}
\keyword{leverage}
\keyword{margin}
\keyword{mark}
\keyword{market}
\keyword{mode.}
\keyword{of}
\keyword{open}
\keyword{order}
\keyword{position}
\keyword{positions}
\keyword{price}
\keyword{recent}
\keyword{settings}
\keyword{single}
\keyword{specific}
\keyword{the}
\keyword{time}
\keyword{timestamp.}
\keyword{under}
\keyword{withdrawal}
\keyword{withdrawals.}
