% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_generator.R
\docType{data}
\name{.gets}
\alias{.gets}
\title{GET request function list}
\format{
An object of class \code{list} of length 123.
}
\usage{
.gets
}
\description{
A list of endpoint functions automatically generated from \code{.api_GET_specs}.
Each function supports \code{tz}, \code{config}, and additional endpoint-specific arguments.
}
\keyword{internal}
