% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_constants.R
\docType{data}
\name{.okx_base_url}
\alias{.okx_base_url}
\title{Base URL for OKX API}
\format{
A character string.
}
\usage{
.okx_base_url
}
\description{
This constant defines the base URL used for all OKX API requests.
}
\keyword{internal}
