% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_get_account.R
\name{get_account_subaccount_max_withdrawal}
\alias{get_account_subaccount_max_withdrawal}
\title{Get sub-account maximum withdrawals}
\usage{
get_account_subaccount_max_withdrawal(
  sub_acct,
  ccy = NULL,
  config,
  tz = .okx_default_tz
)
}
\arguments{
\item{sub_acct}{Character. Sub-account name.}

\item{ccy}{Character or `NULL`. One currency or a comma-separated list of up
to 20 currencies.}

\item{config}{List. API credentials/config.}

\item{tz}{Character. Time zone for parsing timestamps. Default `"Asia/Hong_Kong"`.}
}
\value{
A `data.frame` with per-currency maximum withdrawal values.
}
\description{
Retrieve the maximum withdrawal information for a sub-account from the master
account.
}
