% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_get_asset.R
\name{get_asset_exchange_list}
\alias{get_asset_exchange_list}
\title{Get exchange list}
\usage{
get_asset_exchange_list(tz = .okx_default_tz)
}
\arguments{
\item{tz}{Timezone string.}
}
\value{
A \code{data.frame} with exchange identifiers and names.
}
\description{
Retrieve the public exchange list used by asset withdrawal metadata.
}
