% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_get_trade.R
\name{get_trade_one_click_repay_currency_list_v2}
\alias{get_trade_one_click_repay_currency_list_v2}
\title{Get one-click repay currency list v2}
\usage{
get_trade_one_click_repay_currency_list_v2(config, tz = .okx_default_tz)
}
\arguments{
\item{config}{List. API credentials/config.}

\item{tz}{Character. Time zone for parsing timestamps.}
}
\value{
A `data.frame` with debt and repay currency metadata.
}
\description{
Retrieve repayable currencies for the new one-click repay endpoint.
}
