% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_get_trade.R
\name{get_trade_order_algo}
\alias{get_trade_order_algo}
\title{Get a single algo order}
\usage{
get_trade_order_algo(
  algo_id = NULL,
  algo_cl_ord_id = NULL,
  config,
  tz = .okx_default_tz
)
}
\arguments{
\item{algo_id}{Character or `NULL`. Algo order ID.}

\item{algo_cl_ord_id}{Character or `NULL`. Client algo order ID.}

\item{config}{List. API credentials/config.}

\item{tz}{Character. Time zone for parsing timestamps.}
}
\value{
A `data.frame` with algo order details.
}
\description{
Retrieve a specific algo order by algo ID or client algo order ID.
}
