% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_configure.R
\name{post_account_mmp_config}
\alias{post_account_mmp_config}
\title{Configure MMP}
\usage{
post_account_mmp_config(
  inst_family,
  time_interval,
  frozen_interval,
  qty_limit,
  tz = .okx_default_tz,
  config
)
}
\arguments{
\item{inst_family}{Instrument family.}

\item{time_interval}{Time window in milliseconds.}

\item{frozen_interval}{Frozen interval in milliseconds.}

\item{qty_limit}{Quantity limit in number of contracts.}

\item{tz}{Timezone used for any timestamp parsing.}

\item{config}{A list containing API credentials.}
}
\value{
A `data.frame` describing the applied MMP configuration.
}
\description{
Set market maker protection thresholds for an options instrument family.
}
