% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_configure.R
\name{post_account_move_positions}
\alias{post_account_move_positions}
\title{Move Positions Between Accounts}
\usage{
post_account_move_positions(
  from_acct,
  to_acct,
  legs,
  client_id,
  tz = .okx_default_tz,
  config
)
}
\arguments{
\item{from_acct}{Source account name.}

\item{to_acct}{Destination account name.}

\item{legs}{List of move-position leg objects in the documented OKX shape.}

\item{client_id}{Client-supplied request ID.}

\item{tz}{Timezone used for any timestamp parsing.}

\item{config}{A list containing API credentials.}
}
\value{
A `data.frame` describing the move-position request result.
}
\description{
Move positions between accounts under the same master account.
}
