% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_configure.R
\name{post_account_set_account_level}
\alias{post_account_set_account_level}
\title{Set Account Level}
\usage{
post_account_set_account_level(
  acct_lv,
  tz = .okx_default_tz,
  config
)
}
\arguments{
\item{acct_lv}{Account level string, such as `"1"`, `"2"`, `"3"`, or `"4"`.}

\item{tz}{Timezone used for any timestamp parsing.}

\item{config}{A list containing API credentials.}
}
\value{
A `data.frame` confirming the applied account level.
}
\description{
Switch the account mode.
}
