% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_configure.R
\name{post_account_set_auto_repay}
\alias{post_account_set_auto_repay}
\title{Set Account Auto Repay}
\usage{
post_account_set_auto_repay(
  auto_repay,
  tz = .okx_default_tz,
  config
)
}
\arguments{
\item{auto_repay}{Logical. Whether auto repay should be enabled.}

\item{tz}{Timezone used for any timestamp parsing.}

\item{config}{A list containing API credentials.}
}
\value{
A `data.frame` confirming the applied auto-repay setting.
}
\description{
Enable or disable spot-mode auto repay.
}
