% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_configure.R
\name{post_account_set_greeks}
\alias{post_account_set_greeks}
\title{Set Account Greeks Display Type}
\usage{
post_account_set_greeks(greeks_type, tz = .okx_default_tz, config)
}
\arguments{
\item{greeks_type}{Greeks display type, typically `"PA"` or `"BS"`.}

\item{tz}{Timezone used for any timestamp parsing.}

\item{config}{A list containing API credentials.}
}
\value{
A `data.frame` confirming the applied Greeks display type.
}
\description{
Configure whether Greeks are displayed in PA or BS mode.
}
