% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_configure.R
\name{post_account_set_position_mode}
\alias{post_account_set_position_mode}
\title{Set Account Position Mode}
\usage{
post_account_set_position_mode(
  pos_mode,
  tz = .okx_default_tz,
  config
)
}
\arguments{
\item{pos_mode}{Position mode. Use `long_short_mode` or `net_mode`.}

\item{tz}{Timezone used for any timestamp parsing.}

\item{config}{A list containing API credentials.}
}
\value{
A `data.frame` confirming the applied position mode.
}
\description{
Set the account position mode.
}
