% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_configure.R
\name{post_account_spot_manual_borrow_repay}
\alias{post_account_spot_manual_borrow_repay}
\title{Submit Spot Manual Borrow or Repay}
\usage{
post_account_spot_manual_borrow_repay(
  ccy,
  side,
  amt,
  tz = .okx_default_tz,
  config
)
}
\arguments{
\item{ccy}{Currency.}

\item{side}{Action side, typically `"borrow"` or `"repay"`.}

\item{amt}{Amount.}

\item{tz}{Timezone used for any timestamp parsing.}

\item{config}{A list containing API credentials.}
}
\value{
A `data.frame` describing the executed borrow/repay request.
}
\description{
Manually borrow or repay under spot mode.
}
