% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_asset.R
\name{post_asset_cancel_withdrawal}
\alias{post_asset_cancel_withdrawal}
\title{Cancel an Asset Withdrawal}
\usage{
post_asset_cancel_withdrawal(wd_id, tz = .okx_default_tz, config)
}
\arguments{
\item{wd_id}{Withdrawal request ID.}

\item{tz}{Timezone used for any timestamp parsing.}

\item{config}{API credential list.}
}
\value{
A `data.frame` confirming the cancelled withdrawal ID.
}
\description{
Cancel a pending withdrawal request.
}
