% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_asset.R
\name{post_asset_convert_estimate_quote}
\alias{post_asset_convert_estimate_quote}
\title{Estimate an Asset Convert Quote}
\usage{
post_asset_convert_estimate_quote(
  base_ccy,
  quote_ccy,
  side,
  rfq_sz,
  rfq_sz_ccy,
  cl_q_req_id = NULL,
  tag = NULL,
  convert_mode = NULL,
  tz = .okx_default_tz,
  config
)
}
\arguments{
\item{base_ccy}{Base currency.}

\item{quote_ccy}{Quote currency.}

\item{side}{Quote side, such as `"buy"` or `"sell"`.}

\item{rfq_sz}{RFQ size.}

\item{rfq_sz_ccy}{Currency in which `rfq_sz` is specified.}

\item{cl_q_req_id}{Optional client quote request ID.}

\item{tag}{Optional request tag.}

\item{convert_mode}{Optional OKX convert mode.}

\item{tz}{Timezone used for any timestamp parsing.}

\item{config}{API credential list.}
}
\value{
A `data.frame` describing the estimated conversion quote.
}
\description{
Request a quote for an asset conversion without executing the trade.
}
