% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_trade.R
\name{post_trade_amend_batch_orders}
\alias{post_trade_amend_batch_orders}
\title{Amend Multiple Trade Orders}
\usage{
post_trade_amend_batch_orders(
  orders,
  config,
  tz = .okx_default_tz
)
}
\arguments{
\item{orders}{List of amendment specs.
See \code{post_trade_amend_order()}.}

\item{config}{A list with API credentials.}

\item{tz}{Timezone for parsing response timestamps.}
}
\value{
A `data.frame` with one row per amendment result.
}
\description{
Submit multiple amendment requests in one request.
}
