% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_post_trade.R
\name{post_trade_batch_orders}
\alias{post_trade_batch_orders}
\title{Place Multiple Trade Orders}
\usage{
post_trade_batch_orders(orders, config, tz = .okx_default_tz)
}
\arguments{
\item{orders}{List of order specs.
See \code{post_trade_order()}.}

\item{config}{A list with API credentials.}

\item{tz}{Timezone for parsing response timestamps.}
}
\value{
A `data.frame` with one row per submitted order result.
}
\description{
Submit multiple trade orders in one request.
}
