% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_post_generator.R
\docType{data}
\name{.posts}
\alias{.posts}
\title{POST request function list}
\format{
An object of class \code{list} of length 4.
}
\usage{
.posts
}
\description{
A list of POST endpoint functions generated from \code{.api_POST_specs}.
Each function has the form \code{(body_list, tz, config)} and returns a parsed result.
}
\keyword{internal}
