% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_constants.R
\name{set_okxr_options}
\alias{set_okxr_options}
\title{Set or get okxr options}
\usage{
set_okxr_options(raw_data = NULL)
}
\arguments{
\item{raw_data}{Logical. Whether functions should return raw data (default = FALSE).
If NULL, the function simply returns the current value.}
}
\description{
Convenience wrapper to set global options for okxr, such as whether to
return raw data instead of parsed data.
}
\examples{
okxr_set_options(raw_data = TRUE)
getOption("okxr.raw_data")

set_okxr_options(raw_data = NULL)  # check current value
}
